+++ system/sysauth/user.go
@@ -8,18 +8,18 @@ import (
 	"time"
 )
 
-// User is meant to be a dumb representation of the Skiller struct
-// used in the business logic for dealing with user-based interactions.
+// User is DTO for client views. It is a dumb representation of a Skiller
+// with permissions and roles attached for client logic.
 type User struct {
 	Id       uuid.UUID `json:"id"`
-	Username string    `json:"username" binding:"required"`
-	Password string    `json:"password,omitempty"`
-	Email    string    `json:"email" binding:"required"`
-	RoleId   uuid.UUID `json:"roleId,omitempty" binding:"required"`
-	RoleName string    `json:"role,omitempty" binding:"required"`
-	JWT      string    `json:"accessToken,omitempty"`
+	Username string    `json:"username"`
+	Email    string    `json:"email"`
+	Role     Role      `json:"role"`
+	JWT      string    `json:"accessToken"`
 }
 
+// RegisterUser is the intake-DTO for user/skiller registration. Currently, allow
+// user to assign a role via an `RoleId` or `RoleName` as both have unique db constraints.
 type RegisterUser struct {
 	Username string    `json:"username" binding:"required"`
 	Password string    `json:"password"  binding:"required"`
@@ -28,98 +28,28 @@ type RegisterUser struct {
 	RoleName string    `json:"roleName"`
 }
 
-type LoginUser struct {
-	Email    string `json:"email"`
-	Username string `json:"username"`
-	Password string `json:"password"  binding:"required"`
-}
-
 // UserInteractor implements the UserRepository for use in signon logic.
 type UserInteractor struct {
 	SkillerInteractor *system.SkillerInteractor
 	Logger            api.Logger
 }
 
-// UserRepository is an overlay repository over Skiller. The actual
-// UserInteractor implements all of these  methods by calling its
-// SkillerInteractor and transforming into User structs.
-//
-// The reason to do this for now is to keep User and Skiller separate
-// without having them both live in the database. So for now, Skiller
-// is the source of truth and the User is the DTO wrapper.
+// UserRepository should just be used for creating and editing individual users
+// which includes system-level bindings. It wraps the SkillerInteractor to create
+// the domain-level-user, but handles registration, permissions, roles, etc.
 type UserRepository interface {
-	FindAll() ([]*User, error)
-	FindById(id uuid.UUID) (User, error)
-	FindByUsername(username string) (User, error)
-	FindByEmail(email string) (User, error)
-	CreateOne(user RegisterUser) (User, error)
+	CreateOne(user RegisterUser) (*User, error)
 }
 
-func (u *User) RemovePassword() {
-	u.Password = ""
-}
-
-func (u *UserInteractor) FindAll() ([]*User, error) {
-	skillers, err := u.SkillerInteractor.FindAll()
+func (u *UserInteractor) ExistsByUsername(username string) bool {
+	_, err := u.SkillerInteractor.FindByUsername(username)
 	if err != nil {
-		return nil, err
+		return false
 	}
-	var users []*User
-	for _, skiller := range skillers {
-		users = append(users, &User{
-			Id:       skiller.Id,
-			Username: skiller.Username,
-			Password: skiller.Password,
-			Email:    skiller.Email,
-			RoleId:   skiller.RoleId,
-		})
-	}
-	return users, nil
+	return true
 }
 
-func (u *UserInteractor) FindById(id uuid.UUID) (User, error) {
-	skiller, err := u.SkillerInteractor.FindById(id)
-	if err != nil {
-		return User{}, err
-	}
-	return User{
-		Id:       skiller.Id,
-		Username: skiller.Username,
-		Password: skiller.Password,
-		Email:    skiller.Email,
-		RoleId:   skiller.RoleId,
-	}, nil
-}
-
-func (u *UserInteractor) FindByUsername(username string) (User, error) {
-	skiller, err := u.SkillerInteractor.FindByUsername(username)
-	if err != nil {
-		return User{}, err
-	}
-	return User{
-		Id:       skiller.Id,
-		Username: skiller.Username,
-		Password: skiller.Password,
-		Email:    skiller.Email,
-		RoleId:   skiller.RoleId,
-	}, nil
-}
-
-func (u *UserInteractor) FindByEmail(email string) (User, error) {
-	skiller, err := u.SkillerInteractor.FindByEmail(email)
-	if err != nil {
-		return User{}, err
-	}
-	return User{
-		Id:       skiller.Id,
-		Username: skiller.Username,
-		Password: skiller.Password,
-		Email:    skiller.Email,
-		RoleId:   skiller.RoleId,
-	}, nil
-}
-
-func (u *UserInteractor) CreateOne(user RegisterUser) (User, error) {
+func (u *UserInteractor) CreateOne(user RegisterUser) (*User, error) {
 	skiller, err := u.SkillerInteractor.CreateOne(core.Skiller{
 		Email:        user.Email,
 		Username:     user.Username,
@@ -129,13 +59,11 @@ func (u *UserInteractor) CreateOne(user RegisterUser) (User, error) {
 		DateModified: time.Time{},
 	})
 	if err != nil {
-		return User{}, err
+		return nil, err
 	}
-	return User{
+	return &User{
 		Id:       skiller.Id,
 		Username: skiller.Username,
-		Password: "",
 		Email:    skiller.Email,
-		RoleId:   skiller.RoleId,
 	}, nil
 }